improve base path calculation; disable generation when phpunit runs for a subset of tests

+++ src/Phpunit/TestRunnerExecutionStartedSubscriber.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace MoveMoveIo\Postmangen;
+namespace MoveMoveIo\Postmangen\Phpunit;
 
+use MoveMoveIo\Postmangen\PostmangenConsts;
 use PHPUnit\Event\TestRunner\ExecutionStarted;
 use PHPUnit\Event\TestRunner\ExecutionStartedSubscriber;
 
@@ -11,15 +12,18 @@ class TestRunnerExecutionStartedSubscriber implements ExecutionStartedSubscriber
 
     public function __construct(string $intermediateDir)
     {
-        $this->intermediateDir = trim($intermediateDir, '/');
+        $this->intermediateDir = rtrim($intermediateDir, '/');
     }
 
     public function notify(ExecutionStarted $event): void
     {
         // cleaning up temporary files in case any left from a previous run
 
-        $prefix = PostmangenConsts::TMP_FILE_PREFIX;
-        $files = glob("$this->intermediateDir/$prefix*.json");
+        if (!is_dir($this->intermediateDir)) {
+            return;
+        }
+
+        $files = glob("$this->intermediateDir/*");
 
         foreach ($files as $file) {
             unlink($file);
